feat(ts): passing 404 and ema pages to ts
