+++ package.json
@@ -26,6 +26,7 @@
     "next": "^11.0.1",
     "nprogress": "^0.2.0",
     "prop-types": "^15.7.2",
+    "raw-body": "^2.4.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-html-parser": "^2.0.2",
