--- src/pages/404.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-// TODO: Improve 404 page
-function Error() {
-  return (
-    <div>
-      <h2>Page Not Found</h2>
-    </div>
-  )
-}
-
-export default Error
