+++ src/pages/404.tsx
@@ -1,0 +1,16 @@
+// Dependencies
+import type { NextPage } from 'next'
+
+// TODO: add styles for the 404 page
+const ErrorPage: NextPage = () => {
+  return (
+    <div>
+      <h1>404</h1>
+      <h2>
+        sorry, but the page was <br /> not found
+      </h2>
+    </div>
+  )
+}
+
+export default ErrorPage
